Moved Map options into MenuActivity
Moved Express Lanes, Travel Times, and GoTo location.
This is in prep from adding favoriting map locations.
