+++ src/main/java/gov/wa/wsdot/mobile/client/AppPlaceHistoryMapper.java
@@ -48,11 +48,12 @@ import gov.wa.wsdot.mobile.client.activities.socialmedia.youtube.YouTubeDetailsP
 import gov.wa.wsdot.mobile.client.activities.socialmedia.youtube.YouTubePlace.YouTubePlaceTokenizer;
 import gov.wa.wsdot.mobile.client.activities.tollrates.TollRatesPlace.TollRatesPlaceTokenizer;
 import gov.wa.wsdot.mobile.client.activities.trafficmap.TrafficMapPlace.TrafficMapPlaceTokenizer;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.expresslanes.SeattleExpressLanesPlace.SeattleExpressLanesPlaceTokenizer;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.location.GoToLocationPlace.GoToLocationPlaceTokenizer;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.TrafficMenuPlace.TrafficMenuPlaceTokenizer;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.expresslanes.SeattleExpressLanesPlace.SeattleExpressLanesPlaceTokenizer;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.location.GoToLocationPlace.GoToLocationPlaceTokenizer;
 import gov.wa.wsdot.mobile.client.activities.trafficmap.trafficincidents.TrafficAlertsPlace.TrafficAlertsPlaceTokenizer;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.traveltimes.TravelTimeDetailsPlace.TravelTimeDetailsPlaceTokenizer;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.traveltimes.TravelTimesPlace.TravelTimesPlaceTokenizer;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.traveltimes.TravelTimeDetailsPlace.TravelTimeDetailsPlaceTokenizer;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.traveltimes.TravelTimesPlace.TravelTimesPlaceTokenizer;
 
 import com.google.gwt.place.shared.PlaceHistoryMapper;
 import com.google.gwt.place.shared.WithTokenizers;
@@ -62,6 +63,7 @@ import com.google.gwt.place.shared.WithTokenizers;
         MountainPassesPlaceTokenizer.class,
         MountainPassDetailsPlaceTokenizer.class,
         TrafficMapPlaceTokenizer.class, TollRatesPlaceTokenizer.class,
+        TrafficMenuPlaceTokenizer.class,
         BorderWaitPlaceTokenizer.class, TravelTimesPlaceTokenizer.class,
         GoToLocationPlaceTokenizer.class, BlogPlaceTokenizer.class,
         FacebookPlaceTokenizer.class, TwitterPlaceTokenizer.class,
