+++ src/main/java/gov/wa/wsdot/mobile/client/PhoneAnimationMapper.java
@@ -48,11 +48,12 @@ import gov.wa.wsdot.mobile.client.activities.socialmedia.youtube.YouTubeDetailsP
 import gov.wa.wsdot.mobile.client.activities.socialmedia.youtube.YouTubePlace;
 import gov.wa.wsdot.mobile.client.activities.tollrates.TollRatesPlace;
 import gov.wa.wsdot.mobile.client.activities.trafficmap.TrafficMapPlace;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.expresslanes.SeattleExpressLanesPlace;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.location.GoToLocationPlace;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.TrafficMenuPlace;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.expresslanes.SeattleExpressLanesPlace;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.location.GoToLocationPlace;
 import gov.wa.wsdot.mobile.client.activities.trafficmap.trafficincidents.TrafficAlertsPlace;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.traveltimes.TravelTimeDetailsPlace;
-import gov.wa.wsdot.mobile.client.activities.trafficmap.traveltimes.TravelTimesPlace;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.traveltimes.TravelTimeDetailsPlace;
+import gov.wa.wsdot.mobile.client.activities.trafficmap.menu.traveltimes.TravelTimesPlace;
 
 import com.google.gwt.place.shared.Place;
 import com.googlecode.mgwt.mvp.client.AnimationMapper;
@@ -154,24 +155,24 @@ public class PhoneAnimationMapper implements AnimationMapper {
 			return Animations.SLIDE_REVERSE;
 		}
 		
-		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof TravelTimesPlace) {
+		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof TrafficMenuPlace) {
 			return Animations.SLIDE_UP;
 		}
 
-		if (oldPlace instanceof TravelTimesPlace && newPlace instanceof TrafficMapPlace) {
+		if (oldPlace instanceof TrafficMenuPlace && newPlace instanceof TrafficMapPlace) {
 			return Animations.SLIDE_UP_REVERSE;
+		}
+
+		if (oldPlace instanceof TravelTimesPlace && newPlace instanceof TrafficMenuPlace) {
+			return Animations.SLIDE_REVERSE;
 		}
 		
 		if (oldPlace instanceof TravelTimeDetailsPlace && newPlace instanceof TravelTimesPlace) {
 			return Animations.SLIDE_REVERSE;
 		}
 		
-		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof GoToLocationPlace) {
-			return Animations.SLIDE_UP;
-		}
-		
-		if (oldPlace instanceof GoToLocationPlace && newPlace instanceof TrafficMapPlace) {
-			return Animations.SLIDE_UP_REVERSE;
+		if (oldPlace instanceof GoToLocationPlace && newPlace instanceof TrafficMenuPlace) {
+			return Animations.SLIDE_REVERSE;
 		}
 		
 		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof CameraPlace) {
@@ -193,13 +194,9 @@ public class PhoneAnimationMapper implements AnimationMapper {
 		if (oldPlace instanceof AlertPlace && newPlace instanceof TrafficMapPlace) {
 			return Animations.FADE_REVERSE;
 		}
-
-		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof SeattleExpressLanesPlace) {
-			return Animations.SLIDE_UP;
-		}
 		
-		if (oldPlace instanceof SeattleExpressLanesPlace && newPlace instanceof TrafficMapPlace) {
-			return Animations.SLIDE_UP_REVERSE;
+		if (oldPlace instanceof SeattleExpressLanesPlace && newPlace instanceof TrafficMenuPlace) {
+			return Animations.SLIDE_REVERSE;
 		}
 
 		if (oldPlace instanceof TrafficMapPlace && newPlace instanceof TrafficAlertsPlace) {
