+++ src/main/java/gov/wa/wsdot/mobile/client/activities/trafficmap/menu/TrafficMenuViewGwtImpl.ui.xml
@@ -1,0 +1,27 @@
+<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
+<ui:UiBinder
+        xmlns:ui="urn:ui:com.google.gwt.uibinder"
+        xmlns:g="urn:import:com.google.gwt.user.client.ui"
+        xmlns:mgwt="urn:import:com.googlecode.mgwt.ui.client.widget"
+        xmlns:wsdot="urn:import:gov.wa.wsdot.mobile.client.widget">
+
+    <ui:with field="app" type="gov.wa.wsdot.mobile.client.css.AppBundle" />
+
+    <mgwt:panel.flex.RootFlexPanel>
+        <mgwt:header.HeaderPanel>
+            <wsdot:button.image.BackImageButton ui:field="backButton" />
+            <mgwt:panel.flex.FlexSpacer ui:field="leftFlexSpacer" />
+            <mgwt:header.HeaderTitle ui:field="heading">Traveler Information</mgwt:header.HeaderTitle>
+            <mgwt:panel.flex.FlexSpacer />
+            <mgwt:panel.flex.FixedSpacer />
+        </mgwt:header.HeaderPanel>
+
+        <mgwt:panel.scroll.ScrollPanel scrollingEnabledX="false" ui:field="scrollPanel" styleName="{app.css.voiceoverPage}">
+            <g:FlowPanel>
+                <mgwt:list.celllist.CellList ui:field="cellList" />
+            </g:FlowPanel>
+        </mgwt:panel.scroll.ScrollPanel>
+
+    </mgwt:panel.flex.RootFlexPanel>
+
+</ui:UiBinder>
