+++ src/main/java/gov/wa/wsdot/mobile/client/activities/trafficmap/menu/location/GoToLocationViewGwtImpl.ui.xml
@@ -2,17 +2,18 @@
 <ui:UiBinder
     xmlns:ui="urn:ui:com.google.gwt.uibinder"
     xmlns:g="urn:import:com.google.gwt.user.client.ui"
-    xmlns:mgwt="urn:import:com.googlecode.mgwt.ui.client.widget">
+	xmlns:mgwt="urn:import:com.googlecode.mgwt.ui.client.widget"
+	xmlns:wsdot="urn:import:gov.wa.wsdot.mobile.client.widget">
     
     <ui:with field="app" type="gov.wa.wsdot.mobile.client.css.AppBundle" />
 
 	<mgwt:panel.flex.RootFlexPanel>
 		<mgwt:header.HeaderPanel>
-            <mgwt:panel.flex.FixedSpacer ui:field="leftFixedSpacer" />
-            <mgwt:panel.flex.FlexSpacer ui:field="leftFlexSpacer" />
-			<mgwt:header.HeaderTitle>GoTo Location</mgwt:header.HeaderTitle>
+			<wsdot:button.image.BackImageButton ui:field="backButton" />
+			<mgwt:panel.flex.FlexSpacer ui:field="leftFlexSpacer" />
+			<mgwt:header.HeaderTitle ui:field="heading">Goto Location</mgwt:header.HeaderTitle>
 			<mgwt:panel.flex.FlexSpacer />
-	        <mgwt:button.Button ui:field="doneButton" text="Done" />
+			<mgwt:panel.flex.FixedSpacer />
 		</mgwt:header.HeaderPanel>
 		
         <mgwt:panel.scroll.ScrollPanel ui:field="scrollPanel" scrollingEnabledX="false">
