+++ src/main/java/gov/wa/wsdot/mobile/client/activities/trafficmap/menu/traveltimes/TravelTimeDetailsViewGwtImpl.ui.xml
