+++ src/main/java/gov/wa/wsdot/mobile/client/activities/trafficmap/menu/traveltimes/TravelTimesCell.java
@@ -16,7 +16,7 @@
  *
  */
 
-package gov.wa.wsdot.mobile.client.activities.trafficmap.traveltimes;
+package gov.wa.wsdot.mobile.client.activities.trafficmap.menu.traveltimes;
 
 import gov.wa.wsdot.mobile.client.css.AppBundle;
