+++ src/main/java/gov/wa/wsdot/mobile/client/activities/trafficmap/menu/traveltimes/TravelTimesViewGwtImpl.ui.xml
@@ -2,7 +2,8 @@
 <ui:UiBinder
     xmlns:ui="urn:ui:com.google.gwt.uibinder"
     xmlns:g="urn:import:com.google.gwt.user.client.ui"
-    xmlns:mgwt="urn:import:com.googlecode.mgwt.ui.client.widget">
+    xmlns:mgwt="urn:import:com.googlecode.mgwt.ui.client.widget"
+    xmlns:wsdot="urn:import:gov.wa.wsdot.mobile.client.widget">
     
     <ui:with field="app" type="gov.wa.wsdot.mobile.client.css.AppBundle" />
 
@@ -24,11 +25,11 @@
     
 	<mgwt:panel.flex.RootFlexPanel>
 		<mgwt:header.HeaderPanel>
-            <mgwt:panel.flex.FixedSpacer ui:field="leftFixedSpacer" />
+            <wsdot:button.image.BackImageButton ui:field="backButton" />
             <mgwt:panel.flex.FlexSpacer ui:field="leftFlexSpacer" />
-			<mgwt:header.HeaderTitle ui:field="heading">Travel Times</mgwt:header.HeaderTitle>
-			<mgwt:panel.flex.FlexSpacer />
-	        <mgwt:button.Button ui:field="doneButton" text="Done" />
+            <mgwt:header.HeaderTitle ui:field="heading">Travel Times</mgwt:header.HeaderTitle>
+            <mgwt:panel.flex.FlexSpacer />
+            <mgwt:panel.flex.FixedSpacer />
 		</mgwt:header.HeaderPanel>
 		
         <mgwt:panel.pull.PullPanel ui:field="pullToRefresh">
